# kernel/src/main/java/dto/ScooterDTO.java
package dto;

import java.math.BigDecimal;
import java.util.UUID;

public record ScooterDTO(UUID id,
                         String serialNumber,
                         String brand,
                         String model,
                         BigDecimal batteryLevel,
                         LocationDTO locationDTO,
                         String status) {

}